Remove button works.
We need to attach to the click event of the remove button, not the submit event.  The submit handler is only for <form> elements.

Also, we can speed up the search for the struck elements by moving the .strike selector to into the jquery search.  We can't search for .item2 and .strike at the same time, but that's okay, because .strike will only ever be applied to things that have .item2 anyway.

Finally, we don't remove the .item2.  We remove the parent() of the .item2 (that is, the <li>).